+++ Entities/Concrete/Entities/General/Machine/General/Density.cs
@@ -11,7 +11,7 @@ namespace Entities.Concrete.Entities.General.Machine.General
     {
         public Guid Id { get; set; }
         public string? Alloy { get; set; }// Alaşım 10500
-        public float? Value { get; set; }//2.705
+        public decimal? Value { get; set; }//2.705
         public DateTime Optime { get; set; }
         public bool IsActive { get; set; }
     }
